+++ RFQ.UI/wwwroot/js/Views/Product.js
@@ -181,19 +181,32 @@ function UpdateProduct() {
     });
 }
 function DeleteProduct(itemId) {
-    var deleteProductUrl = '/Product/DeleteProduct/' + itemId;
-    $.ajax({
-        url: deleteProductUrl,
-        type: "DELETE",
-        dataType: "json",
-        data: JSON.stringify(itemId),
-        success: function (response) {
-            toastr.success("Product Deleted Successfully!");
-            $('#currentPage').val(1);
-            FetchProduct();
-        },
-        error: function (xhr, status, error) {
-            toastr.error("Failed to Delete Product!", "Error");
+    Swal.fire({
+        title: 'Are you sure?',
+        text: "This action cannot be undone!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#d33',
+        cancelButtonColor: '#3085d6',
+        confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+        if (result.isConfirmed) {
+            var deleteProductUrl = '/Product/DeleteProduct/' + itemId;
+
+            $.ajax({
+                url: deleteProductUrl,
+                type: "DELETE",
+                dataType: "json",
+                data: JSON.stringify(itemId),
+                success: function (response) {
+                    toastr.success("Product Deleted Successfully!");
+                    $('#currentPage').val(1);
+                    FetchProduct();
+                },
+                error: function (xhr, status, error) {
+                    toastr.error("Failed to Delete Product!", "Error");
+                }
+            });
         }
     });
 }
