+++ RFQ.UI/wwwroot/js/Views/Vehicle.js
@@ -324,27 +324,40 @@ function UpdateVehicle() {
     });
 }
 function DeleteVehicle(vehicleId) {
-    var deleteVehicle = `/Vehicle/DeleteVehicle/${vehicleId}`;
+    Swal.fire({
+        title: 'Are you sure?',
+        text: "This action cannot be undone!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#d33',
+        cancelButtonColor: '#3085d6',
+        confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+        if (result.isConfirmed) {
+            var deleteVehicle = `/Vehicle/DeleteVehicle/${vehicleId}`;
 
-    $.ajax({
-        url: deleteVehicle,
-        type: "DELETE",
-        dataType: "json",
-        success: function (response) {
-            if (response && response.result === "success") {
-                toastr.success("Vehicle Details Deleted Successfully!");
-                $("#addVehicleDiv").addClass('d-none');
-                $('#currentPage').val(1);
-                FetchVehicleList();
-            } else {
-                toastr.error("Failed to Delete Vehicle Details!", "Error");
-            }
-        },
-        error: function (xhr, status, error) {
-            toastr.error("Failed to Delete Vehicle Details!", "Error");
+            $.ajax({
+                url: deleteVehicle,
+                type: "DELETE",
+                dataType: "json",
+                success: function (response) {
+                    if (response && response.result === "success") {
+                        toastr.success("Vehicle Details Deleted Successfully!");
+                        $("#addVehicleDiv").addClass('d-none');
+                        $('#currentPage').val(1);
+                        FetchVehicleList();
+                    } else {
+                        toastr.error("Failed to Delete Vehicle Details!", "Error");
+                    }
+                },
+                error: function (xhr, status, error) {
+                    toastr.error("Failed to Delete Vehicle Details!", "Error");
+                }
+            });
         }
     });
 }
+
 
 function FetchVehicleList() {
     $("#tableDiv").show();
